requirements.txt
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
psycopg2==2.8.4
pytest==5.4.1